IGNITE-25284 Remove polymorphism from "wrapPointer"
